backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cssutils==2.11.0
currency-symbols==2.0.3
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fonttools==4.52.1
idna==3.7
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
webcolors==1.13
xlrd==2.0.1
xls2xlsx==0.2.0
XlsxWriter==3.2.0
gunicorn==20.1.0
python-dotenv